Configuração e alteração do codigo para o Eslint.
