--- src/Components/Footer/Footer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './Footer.module.css';
-
-const Footer = () => {
-  return (
-    <footer className={styles.footer}>
-      <p>Orgulhosamente desenvolvido por <a href='https://github.com/Lu-Pereira'>Luciana Pereira</a> e  <a href='https://github.com/karinesouza'>Karine Souza</a>.</p>
-    </footer>
-  );
-};
-
-export default Footer;
