+++ src/Components/Header.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import styles from './Header.module.css';
+
+const Header = () => (
+  <header className={styles.header}>
+    <nav className={`${styles.nav} container`}>
+      <Link className={styles.login} to="/">
+        Login / Criar
+      </Link>
+    </nav>
+  </header>
+);
+
+export default Header;
