+++ src/Components/Helper/Head.jsx
@@ -1,10 +1,13 @@
-import React from 'react';
+/* eslint-disable prefer-template */
+import React, { useEffect } from 'react';
 
 const Head = (props) => {
-  React.useEffect(() => {
+  useEffect(() => {
+    // eslint-disable-next-line react/prop-types
     document.title = props.title + ' | Burger';
     document
       .querySelector("meta[name='description']")
+      // eslint-disable-next-line react/prop-types
       .setAttribute('content', props.description || '');
   }, [props]);
