--- src/Components/Home.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import Head from './Helper/Head';
-
-const Home = () => {
-  return (
-    <section className="container mainContainer">
-      <Head
-        title='Home'
-        description="Home do burger-game."
-      />
-    </section>
-  );
-};
-
-export default Home;
