+++ src/Components/Home.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Head from './Helper/Head';
+
+const Home = () => (
+  <section className="container mainContainer">
+    <Head
+      title="Home"
+      description="Home do burger-game."
+    />
+  </section>
+);
+
+export default Home;
