+++ src/Components/Login/LoginForm.jsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import { Link, useHistory } from 'react-router-dom';
 import Input from '../Forms/Input';
 import Button from '../Forms/Button';
-import Error from '../Helper/Error';
 import styles from './LoginForm.module.css';
 import stylesBtn from '../Forms/Button.module.css';
 import Head from '../Helper/Head';
@@ -11,13 +10,12 @@ const LoginForm = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [active, setActive] = useState(false);
-  const direcion = useHistory(); 
+  const direcion = useHistory();
 
   const handleClick = () => {
     setActive(!active);
-  }
+  };
 
-  
   const directMenu = () => {
     direcion.push('/menu');
   };
@@ -31,10 +29,9 @@ const LoginForm = () => {
       body: `email=${email}&password=${password}`,
       method: 'POST',
       headers: {
-        'accept': 'application/json',
+        accept: 'application/json',
         'Content-Type': 'application/x-www-form-urlencoded',
-        "Access-Control-Allow-Origin": "*",
-
+        'Access-Control-Allow-Origin': '*',
       },
     })
       .then((response) => response.json())
@@ -56,13 +53,28 @@ const LoginForm = () => {
       <Head title="Login" />
       <h1 className="title">Login</h1>
       <form className={styles.form}>
-        <Input label="Usuário (e-mail)" type="text" value={email} name="name" onChange={(e) => setEmail(e.target.value)} />
-        <Input label="Senha" type="password" value={password} name="password" onChange={(e) => setPassword(e.target.value)} />
-          <Button type='submit' onClick={(e) => {
+        <Input
+          labelText="Usuário (e-mail)"
+          typeInput="text"
+          nameInput="name"
+          onChangeInput={(e) => setEmail(e.target.value)}
+        />
+        <Input
+          labelText="Senha"
+          typeInput="password"
+          nameInput="password"
+          onChangeInput={(e) => setPassword(e.target.value)}
+        />
+        <Button
+          typeButton="submit"
+          onClickBtn={(e) => {
             e.preventDefault();
             handleAuth();
             handleClick();
-          }}>{ active ? 'Carregando...' : 'Entrar'}</Button>
+          }}
+        >
+          {active ? 'Carregando...' : 'Entrar'}
+        </Button>
       </form>
       <div className={styles.register}>
         <h2 className={styles.subtitle}>Cadastre-se</h2>
