--- src/Components/NotFound.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const NotFound = () => {
-  return (
-    <div className="container mainContainer">
-      <h1 className="title">Erro: 404</h1>
-      <p>Página não encontrada.</p>
-    </div>
-  );
-};
-
-export default NotFound;
