+++ src/index.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import  App  from './App';
+import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
-
